+++ server/src/GameServer/GameLogic/Map/Map.Generation.cs
@@ -6,35 +6,103 @@ public partial class Map
 {
     public void GenerateMap()
     {
-        GenerateWalls();
-        GenerateSupplies();
+        try
+        {
+            GenerateWalls();
+            GenerateSupplies();
+
+            // DEBUG: print the map
+            for (int y = 0; y < Height; y++)
+            {
+                string outputStr = "";
+                for (int x = 0; x < Width; x++)
+                {
+                    outputStr += MapChunk[x, y].IsWall ? "#" : " ";
+                }
+                _logger.Information(outputStr);
+            }
+        }
+        catch (Exception e)
+        {
+            _logger.Error(e, "Failed to generate map");
+        }
     }
 
     public void GenerateSupplies()
     {
-        string[] allAvailableSupplies = new string[]
-        {
-            // Weapons
-            Constant.Names.S686, Constant.Names.M16, Constant.Names.VECTOR, Constant.Names.AWM,
+        List<string> weaponNames =
+        [
+            Constant.Names.S686,
+            Constant.Names.M16,
+            Constant.Names.VECTOR,
+            Constant.Names.AWM
+        ];
 
-            // Medicines
-            Constant.Names.BANDAGE, Constant.Names.FIRST_AID,
+        List<string> medicineNames =
+        [
+            Constant.Names.BANDAGE,
+            Constant.Names.FIRST_AID
+        ];
 
-            // Armors
-            Constant.Names.PRIMARY_ARMOR, Constant.Names.PREMIUM_ARMOR,
+        List<string> armorNames =
+        [
+            Constant.Names.PRIMARY_ARMOR,
+            Constant.Names.PREMIUM_ARMOR
+        ];
 
-            // Bullets will be generated with weapons
-
-            // Grenades
+        List<string> grenadeNames =
+        [
             Constant.Names.GRENADE
+        ];
+
+        Dictionary<string, List<string>> supplyNames = new()
+        {
+            { "weapon", weaponNames },
+            { "medicine", medicineNames },
+            { "armor", armorNames },
+            { "grenade", grenadeNames }
         };
+
+        Dictionary<string, double> allAvailableSupplyProba = new()
+        {
+            { "weapon", _random.NextDouble() },
+            { "medicine", _random.NextDouble() },
+            { "armor", _random.NextDouble() },
+            { "grenade", _random.NextDouble() }
+        };
+
+        // Normalize the probabilities
+        double sum = allAvailableSupplyProba.Values.Sum();
+        foreach (string key in allAvailableSupplyProba.Keys.ToList())
+        {
+            allAvailableSupplyProba[key] /= sum;
+        }
+        for (int i = 1; i < allAvailableSupplyProba.Count; i++)
+        {
+            allAvailableSupplyProba[allAvailableSupplyProba.Keys.ElementAt(i)] += allAvailableSupplyProba[allAvailableSupplyProba.Keys.ElementAt(i - 1)];
+        }
+
+        List<string> allAvailableSupplies = [];
+
+        for (int i = 0; i < 1000; i++)
+        {
+            // Random choose a type of item by its probability
+            double randomValue = _random.NextDouble();
+            string itemType = allAvailableSupplyProba.First(x => x.Value >= randomValue).Key;
+
+            // Random choose a specific item
+            string itemSpecificName = supplyNames[itemType][_random.Next(0, supplyNames[itemType].Count)];
+
+            // Add the item to the list
+            allAvailableSupplies.Add(itemSpecificName);
+        }
 
         // Iterate to generate the desired number of supply points
         for (int i = 0; i < _numSupplyPoints; i++)
         {
             Position nextPosition = GenerateValidPosition();
 
-            string itemSpecificName = allAvailableSupplies[_random.Next(0, allAvailableSupplies.Length)];
+            string itemSpecificName = allAvailableSupplies[_random.Next(0, allAvailableSupplies.Count)];
             IItem.ItemKind itemType = IItem.GetItemKind(itemSpecificName);
 
             (int, int) range = GetItemCountRange(itemSpecificName);
@@ -68,6 +136,17 @@ public partial class Map
         // Clear the map
         Clear();
 
+        for (int i = 0; i < 8; i++)
+        {
+
+            ObstacleShape shape = _randomSquareShapes[_random.Next(0, _randomSquareShapes.Count)];
+            PlaceObstacleShape(shape);
+
+            ObstacleShape shape1 = _longWallShapes[_random.Next(0, _longWallShapes.Count)];
+            PlaceObstacleShape(shape1);
+        }
+
+
         // Iterate _tryTimes and place them on the map
         for (int i = 0; i < _tryTimes; i++)
         {
@@ -99,16 +178,16 @@ public partial class Map
     private bool PlaceObstacleShape(ObstacleShape shape)
     {
         // Randomly select position for the obstacle shape
-        int startX = _random.Next(0, Width - shape.MaxWidth);
-        int startY = _random.Next(0, Height - shape.MaxHeight);
+        int startX = _random.Next(0, Width - shape.MaxWidth + 1);
+        int startY = _random.Next(0, Height - shape.MaxHeight + 1);
 
         // Check if the selected position is valid
         if (IsPositionValid(startX, startY, shape))
         {
             // Place the obstacle shape on the map
-            for (int x = 0; x < shape.MaxWidth; x++)
+            for (int x = 0; x < shape.MaxWidth && startX + x < MapChunk.GetLength(0); x++)
             {
-                for (int y = 0; y < shape.MaxHeight; y++)
+                for (int y = 0; y < shape.MaxHeight && startY + y < MapChunk.GetLength(1); y++)
                 {
                     if (shape.IsSolid(x, y))
                     {
@@ -126,15 +205,15 @@ public partial class Map
     private bool IsPositionValid(int startX, int startY, ObstacleShape shape)
     {
         // Check if the position is within the map boundaries
-        if (startX < 0 || startY < 0 || startX + shape.MaxWidth >= Width || startY + shape.MaxHeight >= Height)
+        if (startX < 0 || startY < 0 || startX >= Width || startY >= Height)
         {
             return false;
         }
 
         // Check if the position overlaps with existing obstacles
-        for (int x = 0; x < shape.MaxWidth; x++)
+        for (int x = 0; x < shape.MaxWidth && startX + x < MapChunk.GetLength(0); x++)
         {
-            for (int y = 0; y < shape.MaxHeight; y++)
+            for (int y = 0; y < shape.MaxHeight && startY + y < MapChunk.GetLength(1); y++)
             {
                 if (MapChunk[startX + x, startY + y] != null && MapChunk[startX + x, startY + y].IsWall)
                 {